requirements.txt
wikipedia==1.4.0
dnspython==1.16.0
Flask==1.1.2
pymongo==3.10.1
Pillow==7.1.1